requirements.txt
google-generativeai
llama-index
openai
cohere
sentence_transformers
pypdf
transformers
tree_sitter
tree_sitter_languages
chromadb
faiss-cpu
qdrant-client